Refactor function definitions and handlers in the code
