+++ src/lib/functions/Definitions/functions.ts
